`rake jsdoc`
Requires JSDOCDIR to be set in the environment, or if you just
so happen to store it in ~/vendor/jsdoc-toolkit like I do, then
good for you.

Currently generates completely empty docs. I'd have to put in some
time to get it to generate real content. One of the few problems
with JS: it's so flexible no one writes it the same way. =\

* vendor/jsdoc-simple 0000000...07774a6 (24):
  > Add link to Screenshots