use uint32 when dup'ing test stories.
doesn't speed up chrome (where DataView is native) much (~0.3s), but
firefox (where DataView is JS atop typed arrays) goes from ~6s to ~3s,
so it's worth the possible future issue of stories whose length is
not a multiple of 4. i can deal with that when i get to it.